Clean up source changes.
