+++ cutils.h
@@ -263,7 +263,7 @@ static inline int64_t min_int64(int64_t a, int64_t b)
 /* WARNING: undefined if a = 0 */
 static force_inline int clz32(unsigned int a)
 {
-#ifdef _MSC_VER
+#ifdef _WIN32
     uint32_t leading_zero = 0;
     /* failing to bit scan is undefined, returning 32 instead */
     return (_BitScanReverse(&leading_zero, a)) ? (31 - leading_zero) : 32;
@@ -275,7 +275,7 @@ static force_inline int clz32(unsigned int a)
 /* WARNING: undefined if a = 0 */
 static force_inline int clz64(uint64_t a)
 {
-#ifdef _MSC_VER
+#ifdef _WIN32
     uint32_t leading_zero = 0;
     /* failing to bit scan is undefined, returning 64 instead */
     return (_BitScanReverse64(&leading_zero, a)) ? (63 - leading_zero) : 32;
@@ -287,7 +287,7 @@ static force_inline int clz64(uint64_t a)
 /* WARNING: undefined if a = 0 */
 static inline int ctz32(unsigned int a)
 {
-#ifdef _MSC_VER
+#ifdef _WIN32
     uint32_t trailing_zero = 0;
     /* failing to bit scan is undefined, returning 32 instead */
     return (_BitScanForward(&trailing_zero, a)) ? trailing_zero : 32;
@@ -299,7 +299,7 @@ static inline int ctz32(unsigned int a)
 /* WARNING: undefined if a = 0 */
 static inline int ctz64(uint64_t a)
 {
-#ifdef _MSC_VER
+#ifdef _WIN32
     uint32_t trailing_zero = 0;
     /* failing to bit scan is undefined, returning 64 instead */
     return (_BitScanForward64(&trailing_zero, a)) ? trailing_zero : 64;
